Move `DmfsStyleProvidersTaskTest` to `at.bitfire.synctools.storage.tasks` and clean up unused imports
