+++ synctools/src/androidTest/kotlin/at/bitfire/synctools/storage/tasks/DmfsStyleProvidersTaskTest.kt
@@ -2,11 +2,12 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  */
 
-package at.bitfire.ical4android
+package at.bitfire.synctools.storage.tasks
 
 import android.os.Build
 import androidx.annotation.CallSuper
 import androidx.test.platform.app.InstrumentationRegistry
+import at.bitfire.ical4android.TaskProvider
 import at.bitfire.synctools.test.GrantPermissionOrSkipRule
 import org.junit.After
 import org.junit.Assert.assertNotNull
